Break account creation spec into 2 specs
**Why**: The account creation spec looked like it was really testing
that the handoff to the SP is successful and that the user verifies the
attributes requested by the SP before handoff. This commit breaks that
spec into 2 separate specs to test those things specifically.
