+++ spec/features/idv/sp_requested_attributes_spec.rb
@@ -1,0 +1,11 @@
+require 'rails_helper'
+
+feature 'sp requested IdV attributes', :idv_job, :email do
+  context 'oidc' do
+    it_behaves_like 'sp requesting attributes', :oidc
+  end
+
+  context 'saml' do
+    it_behaves_like 'sp requesting attributes', :saml
+  end
+end
